Refactor the post grid
Heavily refactors the code for generating the grid of posts at the top
of the page to make it not take up as much space in the main render
method.

This essentially breaks everything having to do with the post grid,
including the grid itself, the Prev and Next buttons, and the filter UI,
into a PostDisplay component. Also creates a PostGrid component that
draws just the grid itself without the buttons, and a Post component
that draws a single box for one post.

The PostDisplay component is responsible for everything involving
navigation between pages of posts and keeping track of which post is
currently selected, so all of the functions related to that are moved
into PostDisplay.

However, the main page also needs to know which component is selected so
that it can use that information to either display extra information or
the About Us info. In order to do this, we also add 2 properties,
`onSelected`, and `onUnselected` that are called when a post is either
selected or unselected, respectively.