Fix not moving to first page when filter changes
The previous commit that refactored PostDisplay by moving some of it
into PostGrid made the change in
5f1627c99d127cc0eca29a49ce08b76ed95620fe break by not moving back to the
first page when someone changes the filters.

This fixes that by moving back to the first page whenever a change in
the number of posts occurs, which should happen whenever one of the
filters change.